Cleaned up Dark Theme for merging; Removed Button for better UX on Mobile
